+++ extensions/raindrop-io/src/types.ts
@@ -45,13 +45,6 @@ export type Group = {
   collections: number[];
 };
 
-export interface Preferences {
-  token: string;
-  useLastCollection?: boolean;
-  additionalItemToDisplayInList?: string;
-  displayDate?: string;
-}
-
 export interface CollectionsResponse {
   result: boolean;
   items: Collection[];